Move `Fee` to `Cardano.Read.Ledger.Tx.Fee`
