+++ lib/primitive/lib/Cardano/Wallet/Primitive/Ledger/Read/Tx.hs
@@ -22,6 +22,9 @@ import Cardano.Read.Ledger.Tx.CollateralInputs
 import Cardano.Read.Ledger.Tx.CollateralOutputs
     ( getEraCollateralOutputs
     )
+import Cardano.Read.Ledger.Tx.Fee
+    ( getEraFee
+    )
 import Cardano.Read.Ledger.Tx.Inputs
     ( getEraInputs
     )
@@ -58,9 +61,6 @@ import Cardano.Wallet.Read
     )
 import Cardano.Wallet.Read.Eras
     ( eraValue
-    )
-import Cardano.Wallet.Read.Tx.Fee
-    ( getEraFee
     )
 import Cardano.Wallet.Read.Tx.Hash
     ( getEraTxHash
