+++ lib/wallet/src/Cardano/Wallet/DB/Store/Transactions/TransactionInfo.hs
@@ -24,6 +24,9 @@ import Cardano.Read.Ledger.Tx.CollateralInputs
 import Cardano.Read.Ledger.Tx.CollateralOutputs
     ( getEraCollateralOutputs
     )
+import Cardano.Read.Ledger.Tx.Fee
+    ( getEraFee
+    )
 import Cardano.Read.Ledger.Tx.Inputs
     ( getEraInputs
     )
@@ -100,9 +103,6 @@ import Cardano.Wallet.Read.Eras
     ( EraValue
     , IsEra
     , applyEraFun
-    )
-import Cardano.Wallet.Read.Tx.Fee
-    ( getEraFee
     )
 import Cardano.Wallet.Read.Tx.Hash
     ( getEraTxHash
